Rename test-utils files according to their roles
