+++ frontend/src/mocks/controllers/auth/createUser.ts
@@ -2,7 +2,7 @@ import { SignUpRequest, SignUpResponse } from 'store/thunks';
 import { collection } from 'mocks/models';
 import { sanitizeUser, UserDocument } from 'mocks/models/user';
 import { auth } from 'mocks/utils';
-import { save } from 'mocks/utils/data';
+import { save } from 'mocks/utils/models';
 import { digestText } from 'mocks/utils/crypto';
 
 export const store = (request: SignUpRequest): SignUpResponse => {
